requirement.txt
numpy==1.26.4
natsort
einops
torchsummary
matplotlib
tqdm
scikit-image
scipy
SimpleITK
nibabel